pyproject.toml
[project]
name = "dice-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic[mcp]>=0.52.0",
    "anyio>=4.0.0",
    "flask>=3.0.0",
    "flask-cors>=6.0.2",
    "flask-sqlalchemy>=3.1.1",
    "mcp>=1.6.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.48",
]
